Rename directories to a more standard pattern
Using lower-case names is more common for python modules. Since we will
have multiple requirement files, it also makes sense to put them in a
directory. The dev requirements can inherit from the base requirements.

Later on we will probably want to split settings files into separate
directories as well, but for now that is not important.