Add dj-database-url for parsing db settings
This is one step towards having a [12-factor app](https://12factor.net).

Demanding that all developers have the same database config is tedious.

An app should be easy to get started with, and having a simple database
config is one part of that.

This should also let us use a DATABASE_URL configuration string both on
CI and when deploying to production.