Remove Tox
Tox is certainly a handy tool, but it's overkill for what we are
doing. We don't need to run our tests on multiple different python
platforms.

We have no need for another layer of abstraction. This is just one
app. We can safely rely on py.test and be happy with that.

That also means we can use things like `dj-database-url` to
automatically pick up a DATABASE_URL environment variable when running
tests on CI, and not having to mess around with special configs for
that (Tox seems to hide all env vars and run everything in a vacuum).