Add base models and project scaffolding
This PR adds some of the base modules and models we need to represent
our application domain. We add a view, some models, admin pages and
other things to enable us to start working in multiple places. Consider
this code a scaffold / examples of where to move the code in the future,
rather than anything finished. :)

- show - Stores information about each spex production and performance.
  Not much information is strictly necessary here, but might be expanded
  upon later to add promotional information, etc.
- ticket - App concerned with payment flow and ordering
- venue - Stores information about the performance locations, such as
  theatres and seating arrangements. Will need a fancy admin interface in
  the future.