Replace Main h1 tag using webfonts with an svg image
This will lower the amount of assets loaded when loading the page, since we
don't have to load the Pergament webfont anymore.

The SVG can also be inlined when rendering the HTML server-side to increase
loading speed even more, since that would remove the need for an additional
roundtrip to the server for fetching the image.