Use webpack loader to load asset bundles
This also adds the stripe.js script to every page. Stripe thinks this is a good
idea because it can apparently somehow help with fraud detection.

I did it mostly because I really like keeping script tags in the head tag and
this was the way I managed to solve that.