Use fewer divs for structure
Now, body just contains a list of divs in a flexbox. This means the css becomes
a bit easier to handle, and as far as I can see, it works in both Firefox and
Chrome, on both small and large screen resolutions. This does _not_ mean it
looks nice on mobile devices yet.