Compile used Roboto font faces into style.css
This uses webpack's url-loader to compile the two Roboto woff2 files into the
output.
I don't really like commiting the fontfiles into the repo, but they are fairly
small, and the compiled css file is still only ~30kb.

I also tried using an npm package for the font files, but those seemed to always
import _all_ Roboto files, not the ones used, which meant that some of the
compiled assets ended up at >500kb, which is definitely too much.