Adds working version of Docker-compose.
The image is build to include all dependencies.
All required source is copied over to the image,
this requires long feedback cycles for front-end... Ideas?

Also adds a script to easely create a superuser. This could be seen as
dangerus but if people are in our container we have problem as it is :P