Modify payment intent if reservation changed
If a user goes back from the payment page and decides to select more,
fewer, or other tickets, the sum we want to charge them will likely
change.

We should make sure to modify the payment intent when this happens, so
that we don't accidentally end up charging the wrong amount.